stego_web_app/backend/requirements.txt
django>=5.0
djangorestframework>=3.14
django-cors-headers>=4.3.1
psycopg2-binary>=2.9
celery>=5.3
redis>=5.0
gunicorn>=21.2
# Helper for processing
numpy
scipy
soundfile
sonictag==0.2.0
pydub>=0.25.1
